Renaming CommonFile to BackendInterface file
Changing testers
Documentation
autosub
VELS_WEB

+++ tasks/implementation/VHDL/_backend_interfaces/ghdl_tester_common.sh
@@ -1,5 +1,5 @@
 ######################################################################################
-# Common file for all testers using ISIM from Xilinx ISE 14.7
+# Common file for all testers using ghdl
 #
 # Copyright (C) 2017 Martin  Mosbeck   <martin.mosbeck@gmx.at>
 #                    Gilbert Markum    < >
@@ -13,8 +13,8 @@
 zero=0
 one=1
 
-#path to support files for common scripts
-support_files_path=$common_path/support_files
+#path to support files for backend_interfaces scripts
+support_files_path=$backend_interfaces_path/support_files
 
 #path to autosub
 autosub_path=$(pwd)
@@ -33,7 +33,7 @@ testbench=${task_name}_tb_${user_id}_Task${task_nr}.vhdl
 #echo "desc_path= $desc_path"
 #echo "task_path= $task_path"
 #echo "autosub_path= $autosub_path"
-#echo "common_path= $common_path"
+#echo "backend_interfaces_path= $backend_interfaces_path"
 #echo "support_files_path=$support_files_path"
 #echo "---------------------------------------"
 
@@ -74,7 +74,7 @@ function prepare_test {
 	# create tmp directory for user if it does not exist
 	if [ ! -d "/tmp/$USER" ]
 	then
-		mkdir /tmp/$USER
+   		mkdir /tmp/$USER
 	fi
 
 	# create file for error messages, which will be sent to user
@@ -83,10 +83,10 @@ function prepare_test {
 	#make sure the error_attachments folder is empty
 	if [ ! -d "error_attachments" ];
 	then
-		mkdir error_attachments
+	mkdir error_attachments
 	else
-		rm -r error_attachments
-		mkdir error_attachments
+	rm -r error_attachments
+	mkdir error_attachments
 	fi
 
 	#------ CHECK AND PREPARE USERFILES ------
@@ -95,11 +95,11 @@ function prepare_test {
 		#check if the user supplied a file
 		if [ ! -f $userfile ]
 		then
-			echo "Error with task ${task_nr}. User ${user_id} did not attach the right file."
+			echo "could not test task ${task_nr}. User ${user_id} did not attach the right file."
 			echo "You did not attach your solution. Please attach the file $userfile" > error_msg
 			exit $FAILURE
 		fi
-		
+
 		# delete comments from the file to allow checks like looking for 'wait'
 		# NOTE: this is not a parse and does not cover 2008 multi line
 		# comments, but should work for most cases
@@ -116,19 +116,6 @@ function prepare_test {
 	do
 		desccp $filename
 	done
-
-	# copy the isim tcl file for testing to user's folder
-	# if a special version exists use it, else use common one
-	if [ -f $task_path/scripts/isim.cmd ]
-	then
-		cp $task_path/scripts/isim.cmd $user_task_path
-	else
-		cp $support_files_path/isim.cmd $user_task_path
-	fi
-
-	# prepare ISE
-	. /opt/Xilinx/14.7/ISE_DS/ISE/.settings64.sh /opt/Xilinx/14.7/ISE_DS/ISE
-	unset LD_LIBRARY_PATH
 }
 
 function taskfiles_analyze {
@@ -139,12 +126,25 @@ function taskfiles_analyze {
 
 	for filename in $extrafiles
 	do
-		vhpcomp $filename
+		ghdl -a --ieee=synopsys $filename
 		RET=$?
 		if [ "$RET" -ne "$zero" ]
 		then
 			echo "Error with task ${task_nr} for user ${user_id} while analyzing $filename"
-			echo "Something went wrong with the task ${task_nr} test generation. This is not your" \
+			echo "Something went wrong with the task ${task_nr} test generation. This is not your " \
+			     "fault. We are working on a solution" > error_msg
+			exit $TASKERROR
+		fi
+	done
+
+	for filename in $extrafiles
+	do
+		ghdl -a --ieee=synopsys $filename
+		RET=$?
+		if [ "$RET" -ne "$zero" ]
+		then
+			echo "Error with task ${task_nr} for user ${user_id} while analyzing $filename"
+			echo "Something went wrong with the task ${task_nr} test generation. This is not your " \
 			     "fault. We are working on a solution" > error_msg
 			exit $TASKERROR
 		fi
@@ -152,27 +152,28 @@ function taskfiles_analyze {
 
 	for filename in $entityfiles
 	do
-		vhpcomp $filename
+		ghdl -a --ieee=synopsys $filename
 		RET=$?
 		if [ "$RET" -ne "$zero" ]
 		then
 			echo "Error with task ${task_nr} for user ${user_id} while analyzing $filename"
-			echo "Something went wrong with the task ${task_nr} test generation. This is not your" \
+			echo "Something went wrong with the task ${task_nr} test generation. This is not your " \
 			     "fault. We are working on a solution" > error_msg
 			exit $TASKERROR
 		fi
 	done
 
-	vhpcomp $testbench
+	ghdl -a --ieee=synopsys $testbench
 	RET=$?
 	if [ "$RET" -ne "$zero" ]
 	then
 		echo "Error with task ${task_nr} for user ${user_id} while analyzing the testbench"
-		echo "Something went wrong with the task ${task_nr} test generation. This is not your"\
+		echo "Something went wrong with the task ${task_nr} test generation. This is not your " \
 		     "fault. We are working on a solution" > error_msg
 		exit $TASKERROR
 	fi
-}
+
+	}
 
 function userfiles_analyze {
 	cd $user_task_path
@@ -181,32 +182,17 @@ function userfiles_analyze {
 	for filename in $userfiles
 	do
 		#this is the file from the user
-		vhpcomp $filename 2> /tmp/$USER/tmp_Task${task_nr}_User${user_id}
+		ghdl -a --ieee=synopsys $filename 2> /tmp/$USER/tmp_Task${task_nr}_User${user_id}_analyze
 		RET=$?
-
-		# check for possible infinite loop warning message which crashes the simulation
-		egrep -oq "Possible infinite loop" /tmp/$USER/tmp_Task${task_nr}_User${user_id}
-		RET_loop=$?
-
-		if [ "$RET_loop" -eq "$zero" ]
-		then
-			echo "Task ${task_nr} possible infinite loop for user ${user_id}!"
-			echo "Your submitted behavior file seems to contain an infinite loop. Do all your"
-			     "processes have a sensitivity list?" > error_msg
-			exit $FAILURE
-		fi
 
 		if [ "$RET" -eq "$zero" ]
 		then
 		   echo "Task ${task_nr} analyze success for user ${user_id}!"
 		else
-			echo "Task ${task_nr} analyze FAILED for user ${user_id}!"
-			echo "Analyzation of your submitted behavior file failed:" > error_msg
-
-			# suppress warnings about non usenglish, ERROR & WARTING >>  error_msg
-			cat /tmp/$USER/tmp_Task${task_nr}_User${user_id} | grep -v usenglish | grep WARNING >> error_msg
-			cat /tmp/$USER/tmp_Task${task_nr}_User${user_id} | grep ERROR >> error_msg
-			exit $FAILURE
+		   echo "Task ${task_nr} analyze FAILED for user ${user_id}!"
+		   echo "Analyzation of your submitted behavior file failed:" > error_msg
+		   cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_analyze >> error_msg
+		   exit $FAILURE
 		fi
 	done
 
@@ -216,20 +202,17 @@ function elaborate {
 	cd $user_task_path
 
 	#------ ELABORATE testbench ------#
-
-	# don't store error output as it will be read from log
-	fuse -top ${task_name}_tb 2> /dev/null
+	ghdl -e --ieee=synopsys ${task_name}_tb 2> /tmp/$USER/tmp_Task${task_nr}_User${user_id}_elaborate
 	RET=$?
 
 	if [ "$RET" -eq "$zero" ]
 	then
-		echo "Task${task_nr} elaboration success for user ${user_id}!"
+		echo "Task ${task_nr} elaboration success for user ${user_id}!"
 	else
-		echo "Task${task_nr} elaboration FAILED for user ${user_id}!"
+		echo "Task ${task_nr} elaboration FAILED for user ${user_id}!"
 		echo "Elaboration with your submitted behavior file failed:" > error_msg
-		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id} | grep -v usenglish | grep WARNING >> error_msg # suppress warnings about non usenglish
-		cat fuse.log | grep WARNING >> error_msg
-		cat fuse.log | grep -A 5 ERROR >> error_msg
+		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_analyze >> error_msg
+		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_elaborate >> error_msg
 		exit $FAILURE
 	fi
 }
@@ -237,23 +220,28 @@ function elaborate {
 function simulate {
 	cd $user_task_path
 
-	# set virtual memory limit to 500 MiB
-	ulimit -v $((500*1024))
+	# add parameter for generating the wave file if the wave file shall be attached
+	if [ "$attach_wave_file" -eq "$one" ]
+	then
+		add_wave_file_parameter="--vcd=signals.vcd"
+	else
+		add_wave_file_parameter=""
+	fi
 
-	# start simulation:
-	timeout $simulation_timeout ./x.exe -tclbatch isim.cmd
+	# start simulation, simulation writes to stdout:
+	timeout $simulation_timeout ghdl -r ${task_name}_tb $add_wave_file_parameter > /tmp/$USER/tmp_Task${task_nr}_User${user_id}_simulate
 	RET_timeout=$?
 
 	# check if simulation timed out:
 	if [ "$RET_timeout" -eq 124 ] # timeout exits 124 if it had to kill the process. Probably the simulation has crashed.
 	then
-		echo "Task${task_nr} simulation timeout for user ${user_id}!"
+		echo "Task ${task_nr} simulation timeout for user ${user_id}!"
 		echo "The simulation of your design timed out. This is not supposed to happen. Check your design." > error_msg
 		exit $FAILURE
 	fi
 
 	# check if simulation reported "Success":
-	egrep -q "Success" isim.log
+	egrep -q "Success" /tmp/$USER/tmp_Task${task_nr}_User${user_id}_simulate
 	RET_success=$?
 	if [ "$RET_success" -eq "$zero" ]
 	then
@@ -273,43 +261,16 @@ function simulate {
 		rm signals.vcd
 	fi
 
-	# check if simulation was stopped due to the "run X ms" restriction from the isim.cmd tcl script
-	egrep -q "INFO: Simulator is stopped." isim.log # returns not 0 on exit due to the "run X ms" restriction
-	RET_run_X_ms_restriction=$?
-	if [ "$RET_run_X_ms_restriction" -ne "$zero" ]
-	then
-		echo "Wrong behavior for task ${task_nr} for user ${user_id}!"
-		echo "Your submitted behavior file does not behave like specified in the task description:" > error_msg
-		if [ "$attach_wave_file" -eq "$one" ]
-		then
-			echo "No continuous signal detected. Please look at the attached wave file to see what signal your entity produces." >> error_msg
-		fi
-		exit $FAILURE
-	fi
-
-	# check for simulation errors:
-	cat isim.log | grep -v Security | egrep -q ERROR # security error messages are license errors
-	RET_simulation_error=$?
-	if [ "$RET_simulation_error" -eq "$zero" ]
-	then
-		echo "Simulation error for task ${task_nr} for user ${user_id}"
-		echo "Your submitted behavior file does not behave like specified in the task description:" > error_msg
-		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id} | grep -v usenglish | grep WARNING >> error_msg # suppress warnings about non usenglish
-		cat fuse.log | grep WARNING >> error_msg
-		cat isim.log | grep -v Security | grep -A 5 ERROR >> error_msg
-	exit $FAILURE
-	fi
-
 	# check for the error message from the testbench:
-	egrep -q § isim.log
+	egrep -q § /tmp/$USER/tmp_Task${task_nr}_User${user_id}_simulate
 	RET_tb_error_message=$?
 	if [ "$RET_tb_error_message" -eq "$zero" ]
 	then
 		echo "Wrong behavior for task ${task_nr} for user ${user_id}"
 		echo "Your submitted behavior file does not behave like specified in the task description:" > error_msg
-		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id} | grep -v usenglish | grep WARNING >> error_msg # suppress warnings about non usenglish
-		cat fuse.log | grep WARNING >> error_msg
-		cat isim.log | awk '/§{/,/}§/' | sed 's/§{//g' | sed 's/}§//g' | sed 's/** Failure://g' | sed 's/\\n/\n/g' >> error_msg
+		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_analyze >> error_msg
+		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_elaborate >> error_msg
+		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_simulate | awk '/§{/,/}§/' | sed 's/.*§{//g' | sed 's/}§//g' | sed 's/** Failure://g' | sed 's/\\n/\n/g' >> error_msg
 		if [ "$attach_wave_file" -eq "$one" ]
 		then
 			echo "Please look at the attached wave file to see what signal(s) your entity produces. Use a viewer like GTKWave" \
@@ -318,8 +279,21 @@ function simulate {
 		exit $FAILURE
 	fi
 
+	# check for simulation errors:
+	cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_simulate | egrep -qi error
+	RET_simulation_error=$?
+	if [ "$RET_simulation_error" -eq "$zero" ]
+	then
+		echo "Simulation error for task ${task_nr} for user ${user_id}"
+		echo "Your submitted behavior file does not behave like specified in the task description:" > error_msg
+		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_analyze >> error_msg
+		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_elaborate >> error_msg
+		cat /tmp/$USER/tmp_Task${task_nr}_User${user_id}_simulate  | grep -i error >> error_msg
+		exit $FAILURE
+	fi
+
 	# catch unhandled errors:
-	echo "Unhandled error for task ${task_nr} for user ${user_id}!"
+	echo "Unhandled error in ghdl_tester_common for task ${task_nr} for user ${user_id}!"
 	echo "Your submitted behavior file does not behave like specified in the task description." > error_msg
 	exit $FAILURE
 }
